+++ node/src/hopper/routing_service.rs
@@ -13,7 +13,7 @@ use crate::sub_lib::proxy_client::{
     ClientResponsePayload_0v1, DnsResolveFailure_0v1, ProxyClientSubs,
 };
 use crate::sub_lib::proxy_server::{ClientRequestPayload_0v1, ProxyServerSubs};
-use crate::sub_lib::stream_handler_pool::TransmitDataMsg;
+use crate::sub_lib::neighbor_stream_handler_pool::TransmitDataMsg;
 use actix::Recipient;
 use masq_lib::logger::Logger;
 use std::borrow::Borrow;
