+++ node/src/neighborhood/mod.rs
@@ -63,7 +63,7 @@ use crate::sub_lib::proxy_server::DEFAULT_MINIMUM_HOP_COUNT;
 use crate::sub_lib::route::Route;
 use crate::sub_lib::route::RouteSegment;
 use crate::sub_lib::set_consuming_wallet_message::SetConsumingWalletMessage;
-use crate::sub_lib::stream_handler_pool::DispatcherNodeQueryResponse;
+use crate::sub_lib::neighbor_stream_handler_pool::DispatcherNodeQueryResponse;
 use crate::sub_lib::utils::{handle_ui_crash_request, NODE_MAILBOX_CAPACITY};
 use crate::sub_lib::versioned_data::VersionedData;
 use crate::sub_lib::wallet::Wallet;
@@ -1547,7 +1547,7 @@ mod tests {
     };
     use crate::sub_lib::neighborhood::{NeighborhoodConfig, DEFAULT_RATE_PACK};
     use crate::sub_lib::peer_actors::PeerActors;
-    use crate::sub_lib::stream_handler_pool::TransmitDataMsg;
+    use crate::sub_lib::neighbor_stream_handler_pool::TransmitDataMsg;
     use crate::sub_lib::versioned_data::VersionedData;
     use crate::test_utils::assert_contains;
     use crate::test_utils::make_meaningless_route;
