+++ node/src/node_test_utils.rs
@@ -13,8 +13,8 @@ use crate::neighbor_stream_handler_pool::NeighborStreamHandlerPoolSubs;
 use crate::stream_messages::*;
 use crate::sub_lib::framer::FramedChunk;
 use crate::sub_lib::framer::Framer;
-use crate::sub_lib::stream_handler_pool::DispatcherNodeQueryResponse;
-use crate::sub_lib::stream_handler_pool::TransmitDataMsg;
+use crate::sub_lib::neighbor_stream_handler_pool::DispatcherNodeQueryResponse;
+use crate::sub_lib::neighbor_stream_handler_pool::TransmitDataMsg;
 use crate::test_utils::recorder::Recorder;
 use actix::Actor;
 use actix::Addr;
