+++ node/src/sub_lib/neighbor_stream_handler_pool.rs
@@ -3,6 +3,7 @@ use crate::sub_lib::dispatcher::Endpoint;
 use crate::sub_lib::neighborhood::NodeQueryResponseMetadata;
 use actix::Message;
 
+// This message can be sent either to a neighboring Node or to the client, but not to the server.
 #[derive(PartialEq, Eq, Debug, Message, Clone)]
 pub struct TransmitDataMsg {
     pub endpoint: Endpoint,
